requirements.txt
future==1.0.0
grpcio==1.71.0
logic2-automation==1.0.7
protobuf==5.29.0
psutil==7.0.0
saleae==0.12.0
